Initial commit of a simple SSL cert generation script
This script uses pyOpenSSL to generate self-signed certs. At the
moment there is only basic functionality and a single workflow.
Create CA -> Generate certs -> Write to file

I use this to create certs to be used for machine to machine
encrypted communication.

The next step is to be able to specify using CA files that were
already created to create a client cert.